requirements.txt
boto3
mss
pynput
pyobjc-framework-Cocoa
pillow
